Moved bus.* to hal folder
