+++ cpp/hal/bus.cpp
