+++ cpp/hal/bus.h
