+++ cpp/test/abstract_controller_test.cpp
@@ -8,7 +8,6 @@
 //---------------------------------------------------------------------------
 
 #include "mocks.h"
-#include "bus.h"
 #include "rascsi_exceptions.h"
 #include "controllers/abstract_controller.h"
